+++ themes/hideable.js
@@ -1,83 +1,122 @@
 (function(win) {
 
-  require("IGCMS", function () {
+  require("IGCMS", function() {
 
-    if(IGCMS.Hideable) return;
+    if(IGCMS.Hideable) return
 
-    var Config = {};
+    var Config = {}
 
-    Config.expand = "►";
-    Config.collapse = "▼";
-    Config.hideableClass = "hideable";
-    Config.hideClass = "hide";
-    Config.noHideClass = "nohide";
-    Config.hiddenClass = "hidden";
-    Config.switchClass = "switch";
-    Config.expandTitle = "Expand";
-    Config.collapseTitle = "Collapse";
+    Config.ns = "hideable"
+    Config.expand = "►"
+    Config.collapse = "▼"
+    Config.expandTitle = "Expand"
+    Config.collapseTitle = "Collapse"
+    Config.hideableHiddenClass = Config.ns + "-hidden"
+    Config.hideableNohideClass = Config.ns + "-nohide"
+    Config.hideableNohideFallbackClass = "nohide"
+    Config.hideClass = Config.ns + "-hide"
+    Config.switchClass = Config.ns + "-switch"
+    Config.noprintClass = "noprint"
+    Config.dataHere = "data-" + Config.ns + "-here"
+    Config.dataBefore = "data-" + Config.ns + "-before"
 
-     var Hideable = function() {
+    var Hideable = function() {
 
-      var
-      inited = false,
-      getHideables = function() {
-        if (document.querySelectorAll) return document.querySelectorAll("." + Config.hideableClass);
-        var hideables = [];
-        var allElements = document.getElementsByTagName("*");
-        for(var i = 0; i < allElements.length; i++) {
-          if(allElements[i].classList.contains(Config.hideableClass)) hideables.push(allElements[i]);
+      var inited = false
+
+      var initHideables = function () {
+        var hideables = document.querySelectorAll('.' + Config.ns)
+        for (var i = 0; i < hideables.length; i++) {
+          var link = initToggleButton(hideables[i])
+          if (link === null) {
+            continue
+          }
+          if (hideables[i].classList.contains(Config.hideableNohideClass) || hideables[i].classList.contains(Config.hideableNohideFallbackClass)) {
+            continue
+          }
+          hideables[i].classList.add(Config.hideableNohideClass)
+          toggleElement(link)
         }
-        return hideables;
-      },
-      toggleHideables = function() {
-        var hideables = getHideables();
-        for(var i = 0; i < hideables.length; i++) {
-          var firstElement = hideables[i].children[0];
-          var link = document.createElement("a");
-          link.href = "Javascript:void(0);";
-          link.title = Config.collapseTitle;
-          link.innerHTML = Config.collapse;
-          link.classList.add(Config.switchClass);
-          link.addEventListener("click", toggle, false);
-          firstElement.innerHTML = " " + firstElement.innerHTML;
-          firstElement.insertBefore(link, firstElement.firstChild);
-          if(hideables[i].classList.contains(Config.noHideClass)) continue;
-          hideables[i].classList.add(Config.noHideClass);
-          toggleElement(link);
-        }
-      },
+      }
 
-      toggle = function(e) {
-        var target = e.target || e.srcElement;
-        toggleElement(target);
-        e.preventDefault();
-      },
+      var createToggleButton = function (value) {
+        var link = document.createElement('a')
+        link.href = 'Javascript:void(0)'
+        var linkButton = document.createElement('span')
+        linkButton.innerHTML = Config.collapse
+        link.appendChild(linkButton)
+        link.title = Config.collapseTitle
+        link.classList.add(Config.switchClass)
+        link.classList.add(Config.noprintClass)
+        link.addEventListener('click', toggle, false)
+        link.innerHTML += ' ' + value
+        return link
+      }
 
-      toggleElement = function(link) {
-        var e = link.parentNode.parentNode;
-        if(e.classList.contains(Config.noHideClass)) {
-          e.classList.remove(Config.noHideClass);
-          e.classList.add(Config.hiddenClass);
-          link.innerHTML = Config.expand;
-          link.title = Config.expandTitle;
-        } else {
-          e.classList.remove(Config.hiddenClass);
-          e.classList.add(Config.noHideClass);
-          link.innerHTML = Config.collapse;
-          link.title = Config.collapseTitle;
-        }
-        for(var i = 0; i < e.childNodes.length; i++) {
-          var ch = e.childNodes[i];
-          if(ch.nodeType != 1) continue;
-          if(ch == link.parentNode) continue;
-          if(ch.classList.contains(Config.noHideClass)) continue;
-          if(ch.classList.contains(Config.hideClass)) {
-            ch.classList.remove(Config.hideClass);
-          } else {
-            ch.classList.add(Config.hideClass);
+      var initToggleButton = function (hideable) {
+        for (var i = hideable.children.length - 1; i >= 0; i--) {
+          var item = hideable.children[i]
+          var here = item.getAttribute(Config.dataHere)
+          var before = item.getAttribute(Config.dataBefore)
+          var button = null
+          var value = ""
+          if (before !== null) {
+            value = (before === Config.dataBefore || before === '') ? item.innerHTML : before
+            button = createToggleButton(value)
+            var el = document.createElement(item.nodeName)
+            el.appendChild(button)
+            item.parentNode.insertBefore(el, item)
+            return button
+          }
+          if (here !== null || i === 0) {
+            value = (here === Config.dataHere || !here) ? item.innerHTML : here
+            button = createToggleButton(value)
+            item.innerHTML = ""
+            item.appendChild(button)
+            return button
           }
         }
-      };
+        return null
+      }
+
+      var toggle = function (e) {
+        var target = e.currentTarget
+        toggleElement(target)
+        e.preventDefault()
+      }
+
+      var toggleElement = function (link) {
+        var e = link.parentNode.parentNode
+        var linkButton = link.children[0]
+        if (e.classList.contains(Config.hideableNohideClass)) {
+          e.classList.remove(Config.hideableNohideClass)
+          e.classList.add(Config.hideableHiddenClass)
+          linkButton.innerHTML = Config.expand
+          link.title = Config.expandTitle
+        } else {
+          e.classList.remove(Config.hideableHiddenClass)
+          e.classList.add(Config.hideableNohideClass)
+          linkButton.innerHTML = Config.collapse
+          link.title = Config.collapseTitle
+        }
+        for (var i = e.childNodes.length - 1; i > 0; i--) {
+          var ch = e.childNodes[i]
+          if (ch.nodeType !== 1) {
+            continue
+          }
+          if (ch.firstChild.classList && ch.firstChild.classList.contains(Config.switchClass)) {
+            return
+          }
+          if (ch.classList.contains(Config.hideableNohideClass)) {
+            continue
+          }
+          if (ch.classList.contains(Config.hideClass)) {
+            ch.classList.remove(Config.hideClass)
+          } else {
+            ch.classList.add(Config.hideClass)
+          }
+        }
+      }
 
       // public
       return {
@@ -85,13 +124,13 @@
           if(inited) return;
           IGCMS.initCfg(Config, cfg);
           var css = '/* hideables.js */'
-            + ' .hide { display: none !important; }'
+            + ' .' + Config.hideClass + ' { display: none !important; }'
+            + ' a.' + Config.switchClass + ' span { font-family: "Emilbus Mono", "Lucida Console", monospace; }'
             + ' a.' + Config.switchClass + ' { text-decoration: none;'
             + ' border: none !important;'
-            + ' font-family: "Emilbus Mono", "Lucida Console", monospace;'
-            + ' font-weight: bold }';
+            + ' font-weight: 600; }';
           IGCMS.appendStyle(css);
-          toggleHideables();
+          initHideables();
           inited = true;
         },
         isInit : function() { return inited; }
@@ -99,6 +138,5 @@
     };
 
     IGCMS.Hideable = new Hideable();
-  })
-
+  });
 })(window);