+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
  - Inputvar supports required attribute
 
+### Changed
+ - Refactor hideable to use data attributes
+
 ## [2.11.0] - 2018-04-12
 ### Added
  - Cache ImageList, DocList and processVariables results to improve performance