+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - Inputvar supports required attribute
 
 ### Changed
+ - Improve scrolltopable UX
  - Improve scrolltop UX
 
 ## [2.11.0] - 2018-04-12