+++ CHANGELOG.md
@@ -6,10 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - Honeypot label
  - Inputvar supports required attribute
 
 ### Changed
  - Eventable supports any element
+ - Refactor hideable to use data attributes
+ - Improve scrolltopable UX
 
 ## [2.11.0] - 2018-04-12
 ### Added