+++ CHANGELOG.md
@@ -7,7 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
  - Service worker uses cache-first (fastest) method
+ - Honeypot label
  - Inputvar supports required attribute
+
+### Changed
+ - Eventable supports any element
+ - Refactor hideable to use data attributes
+ - Improve scrolltopable UX
 
 ## [2.11.0] - 2018-04-12
 ### Added
@@ -351,7 +357,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - System messages from previous request are now marked via class.
 
 [Unreleased]: https://bitbucket.org/igwr/cms/compare/dev..master
-[2.11.0]: https://bitbucket.org/igwr/cms/compare/v2.11.0..v2.10.2
+[2.11.0]: https://bitbucket.org/igwr/cms/compare/v2.11.0..v2.10.7
 [2.10.7]: https://bitbucket.org/igwr/cms/compare/v2.10.7..v2.10.6
 [2.10.6]: https://bitbucket.org/igwr/cms/compare/v2.10.6..v2.10.5
 [2.10.5]: https://bitbucket.org/igwr/cms/compare/v2.10.5..v2.10.4